Just change (source[psz]) to (psz < source.length) in line 82.