@oscxc 2021-05-07
> can you show any benchmarks?

`const rand = (a, b) => {
    return Math.floor(Math.random() * (b - a + 1) + a);
};
let BaseX = require('./basex')('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789');

let buffer = Buffer.alloc(16);
for(let i = 0; i < 16; i++){
    buffer[i] = rand(0, 255);
}
let str = BaseX.encode(buffer);
let buffer2 = BaseX.decode(str);
console.log(buffer.equals(buffer2));  //true


let t1 = new Date().getTime();
for (let i = 0; i < 1000000; i++){
    let test = BaseX.decode(str);
}
let t2 = new Date().getTime();
console.log(t2 - t1);

//cpu i9-9900kf

//use present (source[psz])
//593ms, 554ms, 596ms, 607ms, 605ms, 597ms, 627ms, 627ms, 596ms, 622ms

//change to (psz < source.length)
//546ms, 539ms, 563ms, 541ms, 566ms, 546ms, 569ms, 543ms, 539ms, 537ms

`