@steveluscher 2023-01-27
This makes a ton of sense to me. `source` is asserted to be a string, `psz` is monotonically increasing, and `psz >= source.length` is a good indication that you've run out of characters.

Any objection to merging this?